+++ Runtime/Server/ServerEcommerce.cs
@@ -1,10 +1,12 @@
-// Copyright (c) 2020 - 2023 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2020 - 2024 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
-using System;
-using System.Collections.Generic;
+
 using AccelByte.Core;
 using AccelByte.Models;
+using AccelByte.Utils;
+using System;
+using System.Collections.Generic;
 using UnityEngine.Assertions;
 
 namespace AccelByte.Server
@@ -440,11 +442,17 @@ namespace AccelByte.Server
             , ResultCallback<ItemPagingSlicedResultV2> callback )
         {
             Report.GetFunctionLog(GetType().Name);
-            Assert.IsNotNull(criteria, "Can't get items by criteria; Criteria parameter is null!");
 
             if (!session.IsValid())
             {
                 callback.TryError(ErrorCode.IsNotLoggedIn);
+                return;
+            }
+
+            var error = ApiHelperUtils.CheckForNullOrEmpty(criteria);
+            if (error != null)
+            {
+                callback?.TryError(error);
                 return;
             }